\name{FunKink-package}
\alias{FunKink-package}
\docType{package}
\title{Functional Kink Model Estimation and Testing}
\description{
R package for the article \emph{Statistical Inference for Functional Data Models with Kink-structure Changes}.
It provides estimation and testing tools for
\deqn{Y_i(t) = \beta_1(t)X_i + \beta_2(t)(X_i-\gamma)_+ + Z_i^\top\beta_3(t) + e_i(t).}
}
